+++ extensions/gitlab/src/pipeline_my.tsx
@@ -1,0 +1,5 @@
+import { MyPipelinesList } from "./components/pipeline_my";
+
+export default function MyPipelines() {
+  return <MyPipelinesList />;
+}